% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeCheck.ArgumentNull.R
\name{CodeCheck.ArgumentNull}
\alias{CodeCheck.ArgumentNull}
\title{XXXX}
\usage{
CodeCheck.ArgumentNull(
  path_to_dsPackages = NULL,
  package_type = NULL,
  stopMessage = NULL
)
}
\arguments{
\item{path_to_dsPackages}{XXXX}

\item{package_type}{XXXXX}

\item{stopMessage}{XXXXX}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
