% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeCheck.IfStop.R
\name{CodeCheck.IfStop}
\alias{CodeCheck.IfStop}
\title{XXXX}
\usage{
CodeCheck.IfStop(path_to_dsPackages = NULL, package_type = NULL)
}
\arguments{
\item{path_to_dsPackages}{XXXX}

\item{package_type}{XXXXX}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
