% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDSfunction.R
\name{createDSfunction}
\alias{createDSfunction}
\title{Template Builder for client-side DataSHIELD functions}
\usage{
createDSfunction(
  include_function_information = TRUE,
  include_function_arguments = TRUE,
  include_DS_Connections = TRUE,
  include_DS_Connections_Class = TRUE,
  include_isDefined = TRUE,
  include_checkClass = TRUE,
  include_type_Check = TRUE,
  include_methods_Check = TRUE,
  include_newobj = TRUE,
  include_DS_servercall = TRUE,
  function_name = NULL,
  input_object = NULL,
  object_type = NULL,
  datashield_type = NULL
)
}
\arguments{
\item{include_function_information}{is a logical parameter to include the function information in the new R file.}

\item{include_function_arguments}{is a logical parameter to include the function arguments in the new R file.}

\item{include_DS_Connections}{is a logical parameter to include the DSConnection find part in the new R file.}

\item{include_DS_Connections_Class}{is a logical parameter to include the DSConnection class check in the new R file.}

\item{include_isDefined}{is a logical parameter to include the 'defined in all studies' part in the new R file.}

\item{include_checkClass}{is a logical parameter to include the class check of the input variable in the new R file.}

\item{include_type_Check}{is a logical parameter to include the second part of the class check in the new R file.}

\item{include_methods_Check}{is a logical parameter to include the methods check part in the new R file.}

\item{include_newobj}{is a logical parameter to include the newobj generating part in the new R file.}

\item{include_DS_servercall}{is a logical parameter to include the building of the 'cally' object and the call of the server-side function in the new R file.}

\item{function_name}{is the name of the new DataSHIELD function.}

\item{input_object}{is the name of the input object.}

\item{object_type}{is the class of the input object.}

\item{datashield_type}{represent the typ of DataSHIELD server-side function; either 'assign' or 'aggregate'.}
}
\value{
creates a new R file for programming a DataSHIELD function
}
\description{
The function creates a new basic R file for client-side DataSHIELD functions.
}
\details{
This function operates in two steps. Initially, a template is being built based on the selection of the building blocks. In the second step,
this template is filled with information, which were provided by the user and this new R file is being saved in the R folder of the active package.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
