% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilePathFinder.R
\name{FilePathFinder}
\alias{FilePathFinder}
\title{File Path Finder}
\usage{
FilePathFinder(path = path_to_dsPackages, type = package_type)
}
\arguments{
\item{path}{is the directory path to where the DataSHIELD packages are stored.}

\item{type}{refers to the type of DataSHIELD packages: 'Client' or 'Server'}
}
\value{
a data.frame with the function file name and respective file path of either client- or server-side DataSHIELD functions. Internal functions are omitted for now.
}
\description{
This is an internal function.
}
\details{
This is an internal function to retrieve the file path and function file name of either client or server side DataSHIELD functions.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
