% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_buildingblock.R
\name{find_buildingblock}
\alias{find_buildingblock}
\title{Internal Function for pathing of the building blocks for the template}
\usage{
find_buildingblock(buildingblock_name, package = "DSFunctionCreator")
}
\arguments{
\item{buildingblock_name}{specifies which buildingblock is called.}

\item{package}{is the name of the package where the buildingblocks are included ("DSFunctionCreator").}
}
\value{
the path of the buildingblock.
}
\description{
This is an internal function which will find the correct path for the buildingblocks.
}
\details{
This function adapts the code from the internal function find_template from the usethis::use_template function.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
