% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_testblock.R
\name{find_testblock}
\alias{find_testblock}
\title{Internal Function for pathing of the building blocks for the test template of a function}
\usage{
find_testblock(testblock_name, package = "DSFunctionCreator")
}
\arguments{
\item{testblock_name}{specifies which buildingblock is called.}

\item{package}{is the name of the package where the buildingblocks are included ("DSFunctionCreator").}
}
\value{
the path of the buildingblock.
}
\description{
This is an internal function which will find the correct path for the test buildingblocks.
}
\details{
This function adapts the code from the internal function find_template from the usethis::use_template function.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
