% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.dsFunctionTest.R
\name{init.dsFunctionTest}
\alias{init.dsFunctionTest}
\title{XXXX}
\usage{
init.dsFunctionTest(ds.function_name = NULL)
}
\arguments{
\item{path_to_dsPackages}{XXXX}

\item{package_type}{XXXXX}

\item{stopMessage}{XXXXX}
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
