% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.GitHubActions.R
\name{init.GitHubActions}
\alias{init.GitHubActions}
\title{XXXX}
\usage{
init.GitHubActions()
}
\value{
XXXX
}
\description{
XXXX
}
\details{
XXXX
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
