% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_object_template.R
\name{input_object_template}
\alias{input_object_template}
\title{Object input name}
\usage{
input_object_template(input_object)
}
\arguments{
\item{input_object}{is the name of the input object.}
}
\value{
the name of the primary input object according to good practise agreement from DataSHIELD Developer Team
}
\description{
This is an internal function.
}
\details{
This is an internal function to get the correct object name based on the type of the input object
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
