% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_argumentCleaner.R
\name{internal_argumentCleaner}
\alias{internal_argumentCleaner}
\title{Argument Cleaner}
\usage{
internal_argumentCleaner(Client_Arguments = NULL)
}
\arguments{
\item{Client_Arguments}{is the data.frame input object containing the function file name, path and listed arguments.}
}
\value{
a data.frame containing the function name and arguments of the DataSHIELD client functions.
}
\description{
This is an internal function.
}
\details{
This is an internal function to clean up the arguments from the internal_arguments function.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
