% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_arguments.R
\name{internal_arguments}
\alias{internal_arguments}
\title{Codeline and Argument Finder}
\usage{
internal_arguments(df = NULL, type = NULL)
}
\arguments{
\item{df}{is the data.frame input object containing the function file name and path.}

\item{type}{refers to the type of output: either codelines or arguments of a function}
}
\value{
either codelines (a list) of the DataSHIELD functions or a data.frame containing arguments of the DataSHIELD functions.
}
\description{
This is an internal function.
}
\details{
This is an internal function to retrieve the codelines and function arguments of DataSHIELD client functions.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
