% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_newobj.R
\name{internal_newobj}
\alias{internal_newobj}
\title{Codeline Analysis function}
\usage{
internal_newobj(codelines = NULL, Client_Analysis = NULL)
}
\arguments{
\item{codelines}{refers to a list of codelines from DataSHIELD functions}

\item{Client_Analysis}{refers to a data.frame with information on Arguments and DataSHIELD functions}
}
\value{
a data.frame with the function file name and whether the codecheck is present in the function.
}
\description{
This is an internal function.
}
\details{
This is an internal function to compare the codelines of actual functions with pre-defined standards.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
