% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templateDS.R
\name{use_templateDS}
\alias{use_templateDS}
\title{Template filler for CodeCheck}
\usage{
use_templateDS(
  template,
  data = list(),
  save_as = NULL,
  package = "DSFunctionCreator"
)
}
\arguments{
\item{template}{Path to template file relative to `templates/` directory
within `package`; see details.}

\item{data}{A list of data passed to the template.}

\item{save_as}{A name for the file which is going to be saved.}

\item{package}{Name of the package where the template is found.}
}
\value{
A logical vector indicating if file was modified.
}
\description{
This function is very similar to the usethis original but changes the pathing options of where template will be save.
}
\details{
The function pulls a blank codeline file and fills it with arguments to create a final codeline file which can be used for CodeCheck.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
